Updated endpoints
